app/requirements.txt
pymysql==1.1.1
python-dotenv==1.0.1
tabulate==0.9.0
